requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
google-generativeai
